apps/cif-api/pyproject.toml
[project]
name = "cif-api"
version = "0.1.0"
description = "CIF FastAPI backend"
authors = [
    {name = "Jason",email = "jason@stardance.ai"}
]
requires-python = "^3.11"
dependencies = [
    "fastapi (>=0.135.1,<0.136.0)",
    "uvicorn (>=0.41.0,<0.42.0)",
    "sqlalchemy (>=2.0.48,<3.0.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "alembic (>=1.18.4,<2.0.0)",
    "pydantic-settings (>=2.13.1,<3.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "ruff (>=0.15.5,<0.16.0)",
    "black (>=26.3.1,<27.0.0)"
]
